+++ suitkaise/processing/__init__.py
@@ -45,5 +45,5 @@ __all__ = [
 ]
 
 # Module metadata
-__version__ = "0.4.11b0"
+__version__ = "0.4.12"
 __author__ = "Casey Eddings"